admin: Add UsersWithPaging handler and migrate page-based paging stories
Add a UsersWithPaging MSW handler that returns page/size-based pagination
metadata (totalCount, totalPages, nextPage, previousPage) matching the shape
used by createPagePagingActions.

Migrate two stories that demonstrate this pagination style:

- TablePaging — basic page navigation (next/previous)
- TableFilterPagingSort — filter, sort, and pagination combined server-side
